requirements.txt
streamlit
google-generativeai
python-dotenv
gtts
playsound
pyaudio
opencv-python
pytesseract
super_gradients
tesseract
pyttsx3
pycocotools
transformers
easydict
supervision
#bytetrack
#strongsort
numpy